aw_images: initial commit
This is the initial version of the bot task for cleaning up file pages
of copyrighted images on the Animal Well wiki.  Full documentation of
this bot task is at:

	https://animalwell.wiki.gg/wiki/User:AndrybakBot/Image_copyright

The script aw_images.py was adapted from qc_images.py, which was
written by myself for the Questionable Content wiki.  Source code is
located at:

	https://github.com/rybak/questionablecontent/blob/master/bot/qc_images.py